+++ .env-example
@@ -1,3 +1,4 @@
 DATABASE_URL=postgresql://username:password@domain:5432/dbname?schema=public
 NEXTAUTH_URL=http://localhost:3000
 NEXTAUTH_SECRET=changeme
+NEXT_PUBLIC_APP_URL=http://localhost:3000
