+++ src/utils/trpc.ts
@@ -1,20 +1,87 @@
-import { createReactQueryHooks } from '@trpc/react';
-import type { inferProcedureOutput } from '@trpc/server';
+import { httpBatchLink, loggerLink } from '@trpc/client';
+import { createTRPCNext } from '@trpc/next';
+import type { inferProcedureInput, inferProcedureOutput } from '@trpc/server';
 // ℹ️ Type-only import:
 // https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-8.html#type-only-imports-and-export
-import type { AppRouter } from 'server/routers/_app';
+import type { AppRouter } from 'server/trpc/routers/index';
+import superjson from 'superjson';
+
+function getBaseUrl() {
+  // browser should use relative url
+  if (typeof window !== 'undefined') return '';
+
+  // reference for vercel.com SSR
+  if (process.env.VERCEL_URL) {
+    return `https://${process.env.VERCEL_URL}`;
+  }
+
+  // reference for render.com SSR
+  if (process.env.RENDER_INTERNAL_HOSTNAME) {
+    return `http://${process.env.RENDER_INTERNAL_HOSTNAME}:${process.env.PORT}`;
+  }
+
+  // override for docker etc SSR
+  if (process.env.NEXT_PUBLIC_APP_URL) return process.env.NEXT_PUBLIC_APP_URL;
+
+  // assume localhost in dev SSR
+  return `http://localhost:${process.env.PORT ?? 3000}`; // dev SSR should use localhost
+}
+
+export const trpc = createTRPCNext<AppRouter>({
+  /**
+   * Config options:
+   * https://trpc.io/docs/v10/nextjs#createtrpcnext-options
+   */
+  config() {
+    return {
+      /**
+       * @link https://trpc.io/docs/v10/data-transformers
+       */
+      transformer: superjson,
+      /**
+       * @link https://trpc.io/docs/v10/links
+       */
+      links: [
+        // adds pretty logs to your console in development and logs errors in production
+        loggerLink({
+          enabled: (opts) =>
+            process.env.NODE_ENV === 'development' ||
+            (opts.direction === 'down' && opts.result instanceof Error),
+        }),
+        httpBatchLink({
+          url: `${getBaseUrl()}/api/trpc`,
+        }),
+      ],
+    };
+  },
+  /**
+   * @link https://trpc.io/docs/v10/ssr
+   */
+  ssr: false,
+});
 
 /**
  * A set of strongly-typed React hooks from your `AppRouter` type signature with `createReactQueryHooks`.
  * @link https://trpc.io/docs/react#3-create-trpc-hooks
  */
-export const trpc = createReactQueryHooks<AppRouter>();
 
-// export const transformer = superjson;
 /**
- * This is a helper method to infer the output of a query resolver
+ * These are helper types to infer the input and output of query resolvers
  * @example type HelloOutput = inferQueryOutput<'hello'>
  */
+// TODO: how do these work in v10?
 export type inferQueryOutput<
   TRouteKey extends keyof AppRouter['_def']['queries'],
 > = inferProcedureOutput<AppRouter['_def']['queries'][TRouteKey]>;
+
+export type inferQueryInput<
+  TRouteKey extends keyof AppRouter['_def']['queries'],
+> = inferProcedureInput<AppRouter['_def']['queries'][TRouteKey]>;
+
+export type inferMutationOutput<
+  TRouteKey extends keyof AppRouter['_def']['mutations'],
+> = inferProcedureOutput<AppRouter['_def']['mutations'][TRouteKey]>;
+
+export type inferMutationInput<
+  TRouteKey extends keyof AppRouter['_def']['mutations'],
+> = inferProcedureInput<AppRouter['_def']['mutations'][TRouteKey]>;
